requirements.txt
Flask
gunicorn
pytest
sklearn
pandas
numpy
PYAML

